Output processor tests
