+++ packages/core/src/llm/model/model.loop.ts
@@ -144,8 +144,6 @@ export class MastraLLMVNext extends MastraBase {
       tools: Object.keys(tools || {}),
     });
 
-    console.log('objectOptionsStream', objectOptions);
-
     let schema: z.ZodType<Z> | Schema<Z> | undefined;
     if (experimental_output) {
       this.logger.debug('[LLM] - Using experimental output', {
