+++ packages/core/src/processors/processors/structured-output.ts
@@ -3,6 +3,7 @@ import { Agent } from '../../agent';
 import type { MastraMessageV2 } from '../../agent/message-list';
 import type { StructuredOutputOptions } from '../../agent/types';
 import type { Processor } from '../index';
+import { RuntimeContext } from '../../runtime-context';
 
 export type { StructuredOutputOptions } from '../../agent/types';
 
@@ -58,16 +59,28 @@ export class StructuredOutputProcessor<S extends z.ZodTypeAny> implements Proces
           return message;
         }
 
-        console.log('textContent', textContent)
+        console.log('textContent', textContent);
 
         try {
-          // Use structuring agent to extract structured data from the unstructured text
-          const structuredResult = await this.structuringAgent.generate(
-            `Extract and structure the key information from the following text according to the specified schema. Keep the original meaning and details:\n\n${textContent}`,
-            {
-              output: this.schema,
-            },
-          );
+          const modelDef = await this.structuringAgent.getModel({ runtimeContext: new RuntimeContext() });
+
+          let structuredResult;
+          if (modelDef.specificationVersion === 'v1') {
+            // Use structuring agent to extract structured data from the unstructured text
+            structuredResult = await this.structuringAgent.generate(
+              `Extract and structure the key information from the following text according to the specified schema. Keep the original meaning and details:\n\n${textContent}`,
+              {
+                output: this.schema,
+              },
+            );
+          } else {
+            structuredResult = await this.structuringAgent.generate_vnext(
+              `Extract and structure the key information from the following text according to the specified schema. Keep the original meaning and details:\n\n${textContent}`,
+              {
+                output: this.schema,
+              },
+            );
+          }
 
           if (!structuredResult.object) {
             this.handleError('Structuring failed', 'Internal agent did not generate structured output', abort);
@@ -88,8 +101,6 @@ export class StructuredOutputProcessor<S extends z.ZodTypeAny> implements Proces
 
             return message;
           }
-
-          console.log('structuredResult', structuredResult.object)
 
           // Store both original text and structured data in a way the agent can use
           // The agent expects text but we need both text and object for experimental_output
